LayoutTests/accessibility/mac/search-predicate-visible-button-expected.txt

first blockquote level 1
second blockquote level 1
sans-serif black bold text with underline

serif blue italic text


 checkbox


first heading level 1

second heading level 1

heading level 2

heading level 3

heading level 4

heading level 5

heading level 6

landmark

link
list item
live region

radio button
first table heading level 1
first table body level 1
second table heading level 1
second table body level 1
tree item
isVisible test start

test button 1
test test test test
test test test test
test test test test
test test test test
test test test test
test test test test
test test test test
test test test test
test test test test
test button 2
test button 3
This tests the ability to search for accessible elements by key or text.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS resultElement.role is 'AXRole: AXButton'
PASS resultElement.title is 'AXTitle: Submit'
PASS resultElement is undefined.
PASS successfullyParsed is true

TEST COMPLETE
